Move sentinels to ``sphinx.ext.autodoc._sentinels``
